inverse spatial join to accelerate selection of tiles and add file map
When selecting the tiles for a given chunk, there is one geometry per tile and
one per chunk. Making the chunk element the left dataset in the sjoin() call
seems to make the request much faster (50x on a small test with 8 files
elements over 8000).

I have change the list of files into a dictionary to make mapping from tile
name to file name simpler in the code, and likely faster (although I have not
measured if it was bottleneck).