try to avoid in-place assignement to set missing data values
I try to avoid accessing the low-level .data dask array object, favoring the
xarray api if possible, as well as avoiding potential in-place updates.
It may not make a difference, but it helps (me :D) remember that actually the
underlying operation consist in filtering the whole array, keeping some values
and replacing others (so what we read is closer to what Dask does).